BOLT 04: redfine mpp HTLC set identifier to use payment_secret
The prior definition defined an HTLC set to be dependent on payment
preimage/hash. This commit modifies the definition to use the
payment_secret as the primary identifier.

Due to the uniqueness requirements of invoice preimages and
payment_secrets, this doesn't introduce any meaningful changes for
implementations. However, this does better serve to unify the defintion
between MPP and the upcoming AMP defintion of an HTLC set.